East side west side everybody coming down to meet the M E T S mets of new york town
MEET THE METS
MEET THE METS
Step right up and greet the Mets!
Bring your kiddies
Bring your wife;
Guaranteed to have the time of your life
Because the Mets are really sockin' the ball; knocking those home runs over the wall!
Eastside
Westside
Everybody's coming down
To meet the M-E-T-S Mets of New York town!
Oh, the butcher and the baker and the people on the streets
Where did they go? To MEET THE METS!
Oh, they're hollerin' and cheerin' and they're jumpin' in their seats
Where did they go? To MEET THE METS!
All the fans are true to the orange and blue
So hurry up and come on down -
'Cause we've got ourselves a ball club
The Mets of New York town!
Give 'em a yell!
Give 'em a hand!
And let 'em know your rootin' in the stand!
Come on and MEET THE METS
MEET THE METS
Step right up and greet the Mets!
Bring your kiddies
Bring your wife;
Guaranteed to have the time of your life
Because the Mets are really sockin' the ball; knocking those home runs over the wall!
Eastside
Westside
Everybody's coming down
To meet the M-E-T-S Mets of New York town!
Of New York town!